edit folder dashboard
