+++ capstone-react-9/src/components/AdminPage/index.jsx
@@ -1,0 +1,9 @@
+import DashboardPage from "./DashboardPage/index.jsx"
+
+export default function adminPage() {
+  return (
+    <div className='flex font-Poppins' style={{ backgroundColor: '#E6F6FF' }}>
+        <DashboardPage/>
+    </div>
+  )
+}
